Clean up Replay Lab model matrix
Split Replay Lab internals into focused modules, make replay trust and candidate verdicts canonical, and gate MCP evaluator descriptors behind an explicit trusted-local opt-in.\n\nClarify the model-matrix docs around production decision evidence, separate the observed current alias from replacement candidates, and lock the sample report to the renderer output.
