+++ examples/end_to_end/replay_lab/langgraph_requirements_triage/README.md
@@ -1,10 +1,10 @@
 # LangGraph requirements-triage model matrix
 
-This is the live Replay Lab example for comparing several model aliases on the same LangGraph agent.
+A product team already has a requirements-triage agent in production. The agent reads messy requests, summarizes the ask, lists known requirements, points out missing information, names risks, and recommends the next action.
 
-The example is intentionally small. A product team brings a few messy requirements requests to an assistant. The assistant is not supposed to solve the whole product problem. Its job is to triage the request: summarize it, list known requirements, point out missing information, name risks, and recommend the next action.
+Later, the team wants to change something: maybe a cheaper model, a faster model, or a prompt/model combination that should preserve the same behavior. The danger is not dramatic failure. The danger is boring, easy-to-miss failure: the answer still sounds fluent, but it drops the missing approval rule; it keeps the headings, but stops warning about deadline risk.
 
-That gives Replay Lab something concrete to measure. Model swaps often fail in boring, easy-to-miss ways. A cheaper model might sound fluent but skip the missing approval rule. A faster model might keep the right headings but stop warning about risk. Replay Lab replays the same cases so you can compare those differences side by side.
+This live Replay Lab example is about gathering evidence for that decision. It replays the same requirements cases through candidate model aliases so you can inspect whether a replacement preserves the behavior that matters and whether any cost or latency savings are worth the risk. It is not a universal model leaderboard.
 
 ## What is inside
 
@@ -30,7 +30,7 @@ The main files are:
 - `requirements_cases.py`: three synthetic requirements requests, with stable case IDs and labels.
 - `requirements_flow.py`: the live Kitaru + LangGraph flow. It uses the runner name `requirements_triage`, so the replay anchor is `requirements_triage_langgraph_call`.
 - `seed_observed.py`: runs the live flow to create observed executions, then writes a Replay Lab manifest.
-- `candidates/model_matrix.example.json`: three candidate descriptors using model aliases: `cheap`, `balanced`, and `quality`.
+- `candidates/model_matrix.example.json`: three replacement-candidate descriptors using model aliases: `cheap`, `balanced`, and `quality`.
 - `run_replay_lab.py`: replays the manifest against the candidate matrix and writes JSON/Markdown reports.
 - `evaluator.py`: a deterministic evaluator that checks the answer shape. It looks for the expected sections and false-certainty language.
 - `render_report.py`: turns a Replay Lab JSON report into a single HTML report.
@@ -42,11 +42,11 @@ Each case has three kinds of lanes:
 
 1. **Observed**: the original live run created by `seed_observed.py`.
 2. **Baseline replay**: the same execution replayed from `requirements_triage_langgraph_call` with no candidate change.
-3. **Candidate replay**: the same replay point, but with a candidate model alias passed in.
+3. **Candidate replay**: the same replay point, but with a replacement-candidate model alias passed in.
 
 The baseline lane is the control. If replay alone changes the answer, cost, or quality, the report lowers its confidence. The candidate lanes are then judged against that baseline, not against the original observed run directly.
 
-In plain terms, Replay Lab first asks: did replay itself move the target? Then it asks: what did each candidate change beyond that?
+In plain terms, Replay Lab first asks: did replay itself move the target? Then it asks: what did each proposed replacement change beyond that?
 
 ## Before running the live path
 
@@ -58,9 +58,10 @@ uv run kitaru login
 uv run kitaru init
 ```
 
-The committed files use model aliases, not provider model names. Register aliases that make sense for your own environment:
+The committed files use model aliases, not provider model names. Register aliases that make sense for your own environment. Use `current` for the model that produced the observed lane, and use the other aliases for possible replacements:
 
 ```bash
+uv run kitaru model register current --model <provider/model-for-current-production-alias>
 uv run kitaru model register cheap --model <provider/model-for-cheap-alias>
 uv run kitaru model register balanced --model <provider/model-for-balanced-alias>
 uv run kitaru model register quality --model <provider/model-for-quality-alias>
@@ -69,8 +70,8 @@ uv run kitaru model list
 
 For the first live run, you need at least:
 
-- `balanced` for observed seeding
-- `cheap` and `balanced` for the first two-candidate comparison
+- `current` for observed seeding
+- `cheap` and `balanced` for the first two-candidate replacement comparison
 
 ## First: inspect the sample report
 
@@ -92,7 +93,7 @@ Open `/tmp/requirements-triage-sample.html`. Look for:
 
 - the overall recommendation
 - replay trust
-- candidate ranking
+- candidate decision evidence
 - changed output sections
 - evaluator details
 - cases the report says to inspect
@@ -106,7 +107,7 @@ Seed two observed live runs:
 ```bash
 uv run python examples/end_to_end/replay_lab/langgraph_requirements_triage/seed_observed.py \
   --small \
-  --model balanced
+  --model current
 ```
 
 This writes:
@@ -138,7 +139,7 @@ Then open:
 reports/requirements-triage-langgraph-demo.html
 ```
 
-If that run looks good, remove `--candidate-limit 2` to include the `quality` alias as well.
+If that run looks good, remove `--candidate-limit 2` to include the `quality` replacement candidate as well.
 
 ## Useful command variants
 
@@ -146,7 +147,7 @@ Seed all three committed cases:
 
 ```bash
 uv run python examples/end_to_end/replay_lab/langgraph_requirements_triage/seed_observed.py \
-  --model balanced
+  --model current
 ```
 
 Seed a specific case:
@@ -154,7 +155,7 @@ Seed a specific case:
 ```bash
 uv run python examples/end_to_end/replay_lab/langgraph_requirements_triage/seed_observed.py \
   --case onboarding-workflow-access \
-  --model balanced
+  --model current
 ```
 
 Use repeated candidate files instead of the matrix file:
@@ -170,13 +171,13 @@ Do not pass `--matrix-path` and `--candidate-path` together. The script rejects 
 
 ## Reading the verdict
 
-The HTML report uses three verdict words:
+The HTML report uses three verdict words. Read them as decision evidence for this replay cohort, not as a universal ranking of models:
 
 - `ship` means the candidate looks safe enough on this cohort to consider moving forward.
 - `caution` means the candidate may be useful, but the report found cases you should read before deciding.
 - `hold` means the comparison is not strong enough shipping evidence.
 
-The most important warning is replay drift. If baseline replay already changes the answer too much before a candidate is applied, the candidate result is weaker evidence. The report can still be useful, but it is telling you to inspect the case instead of trusting the ranking blindly.
+The most important warning is replay drift. If baseline replay already changes the answer too much before a candidate is applied, the candidate result is weaker evidence. The report can still be useful, but it is telling you to inspect the case instead of treating the ordered candidates as a winner board.
 
 ## Generated files and git hygiene
