+++ examples/end_to_end/replay_lab/langgraph_requirements_triage/evaluator.py
@@ -44,6 +44,7 @@ def evaluate_requirements_triage(request: Any) -> dict[str, Any]:
     return {
         "quality_score": quality_score,
         "scorecard": scorecard,
+        "drift_signature": scorecard,
         "limitations": limitations,
     }
